+++ Word2Vec.Net/BestWord.cs
@@ -1,0 +1,8 @@
+namespace Word2Vec.Net
+{
+  public struct BestWord
+  {
+    public string Word { get; set; }
+    public float Distance { get; set; }
+  }
+}
