+++ Word2Vec.Net/Distance.cs
@@ -93,10 +93,4 @@ namespace Word2Vec.Net
 
 
     }
-
-    public struct BestWord
-    {
-        public string Word { get; set; }
-        public float Distance { get; set; }
-    }
 }
